SVY-7664 Related records are wrongfully deleted when having a duplicate index

+++ servoy_shared/src/com/servoy/j2db/dataprocessing/FoundSet.java
@@ -3968,19 +3968,23 @@ public abstract class FoundSet implements IFoundSetInternal, IRowListener, Scrip
 						}
 					}
 				}
-
-				// delete the related data
-				it = fsm.getApplication().getFlattenedSolution().getRelations(sheet.getTable(), true, false);
-				while (it.hasNext())
+				//if this is a new record, this delete operation may be the result of a rollback due to a duplicated pk
+				// so only delete related data if the record is already in the db
+				if (state.existInDataSource())
 				{
-					Relation rel = it.next();
-					if (rel.getDeleteRelatedRecords() && !rel.isGlobal())//if completely global never delete do cascade delete
+					// delete the related data
+					it = fsm.getApplication().getFlattenedSolution().getRelations(sheet.getTable(), true, false);
+					while (it.hasNext())
 					{
-						IFoundSetInternal set = state.getRelatedFoundSet(rel.getName());
-						if (set != null)
+						Relation rel = it.next();
+						if (rel.getDeleteRelatedRecords() && !rel.isGlobal())//if completely global never delete do cascade delete
 						{
-							Debug.trace("******************************* delete related set size: " + set.getSize() + " from record with PK: " + state.getPKHashKey() + " index in foundset: " + row); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
-							set.deleteAllInternal();
+							IFoundSetInternal set = state.getRelatedFoundSet(rel.getName());
+							if (set != null)
+							{
+								Debug.trace("******************************* delete related set size: " + set.getSize() + " from record with PK: " + state.getPKHashKey() + " index in foundset: " + row); //$NON-NLS-1$ //$NON-NLS-2$ //$NON-NLS-3$
+								set.deleteAllInternal();
+							}
 						}
 					}
 				}
@@ -5125,7 +5129,7 @@ public abstract class FoundSet implements IFoundSetInternal, IRowListener, Scrip
 
 	/*
 	 * (non-Javadoc)
-	 * 
+	 *
 	 * @see com.servoy.j2db.dataprocessing.IFoundSetInternal#setSelectedIndex(java.lang.String)
 	 */
 	@Override
@@ -5877,7 +5881,7 @@ public abstract class FoundSet implements IFoundSetInternal, IRowListener, Scrip
 
 	/*
 	 * (non-Javadoc)
-	 * 
+	 *
 	 * @see java.lang.Object#toString()
 	 */
 	@Override
